@@ -5,19 +5,13 @@ import 'package:flutter_test/flutter_test.dart';
 void main() {
   Widget makeTestableWidget({Widget child}) {
     return MaterialApp(
-        theme: ThemeData(primaryColor: Colors.pink,
-        brightness: Brightness.light),
-        home: Scaffold(
-            body: Center(),
-            bottomNavigationBar: child));
+        theme: ThemeData(primaryColor: Colors.pink, brightness: Brightness.light),
+        home: Scaffold(body: Center(), bottomNavigationBar: child));
   }
 
   testWidgets('Fancy Nav has correct tabs', (WidgetTester tester) async {
     FancyBottomNavigation fn = FancyBottomNavigation(
-      tabs: [
-        TabData(iconData: Icons.home, title: "Home"),
-        TabData(iconData: Icons.search, title: "Search")
-      ],
+      tabs: [TabData(icon: Icons.home, title: "Home"), TabData(icon: Icons.search, title: "Search")],
       onTabChangedListener: (position) {},
     );
 
@@ -37,15 +31,11 @@ void main() {
 
     final randomFinder = find.text("Hello");
     expect(randomFinder, findsNothing);
-
   });
 
   testWidgets('Clicking icon moves the circle', (WidgetTester tester) async {
     FancyBottomNavigation fn = FancyBottomNavigation(
-      tabs: [
-        TabData(iconData: Icons.home, title: "Home"),
-        TabData(iconData: Icons.search, title: "Search")
-      ],
+      tabs: [TabData(icon: Icons.home, title: "Home"), TabData(icon: Icons.search, title: "Search")],
       onTabChangedListener: (position) {},
     );
 
@@ -68,7 +58,5 @@ void main() {
 
     expect(searchIconFinder, findsNWidgets(2));
     expect(homeIconFinder, findsOneWidget);
-
   });
-
 }
